feat(algo): harden rknn dma-buf cache and add package-scoped env tuning
- Harden RKNN DMA-BUF zero-copy pipeline with bounded LRU cache (16),
  strict fstat validation, and paired RAII unmap/destroy across
  rk3568/rk3576 algorithm packages.
- Flatten face recognition config schemas for frontend rendering with
  backward-compatible deserialization and fix cross-compile Makefiles.
- Introduce PackageEnv in algo-sdk for in-memory, package-scoped .env
  parsing with zero global process environment leakage.
- Implement three-tier configuration precedence hierarchy:
  host task.parameters > package-scoped .env > hardcoded defaults.
- Decouple model paths from mandatory manifest.json declarations,
  annotate all .env.example templates, and update Nuwa algo guidelines.

+++ algo-packages/rknn/rk3576/face_recognition/src/rknn.rs
@@ -1242,7 +1242,7 @@ fn dma_identity(layout: &DmaBufLayout) -> Result<DmaIdentity, AlgoError> {
     // SAFETY: fstat 返回成功后完整写入 stat。
     let stat = unsafe { stat.assume_init() };
     Ok(DmaIdentity {
-        device: stat.st_dev,
+        device: stat.st_dev as u64,
         inode: stat.st_ino,
         size: layout.size,
         stride: layout.stride[0],
